requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
absl-py==0.7.1            # via tensorboard, tensorflow-estimator, tensorflow-gpu
appdirs==1.4.3            # via black, dvc
asciimatics==1.10.0       # via dvc
astor==0.7.1              # via tensorflow-gpu
attrs==19.1.0             # via black
bert-tensorflow==1.0.1
black==19.3b0
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0                # via black, pip-tools
colorama==0.4.1           # via dvc
configobj==5.0.6          # via dvc
configparser==3.7.4       # via dvc
contextlib2==0.5.5        # via schema
decorator==4.4.0          # via jsonpath-rw, networkx
distro==1.4.0             # via dvc
dvc==0.35.7
fastparquet==0.3.1
future==0.17.1            # via asciimatics, dvc, grandalf
gast==0.2.2               # via tensorflow-gpu
gitdb2==2.0.5             # via gitpython
gitpython==2.1.11         # via dvc
grandalf==0.6             # via dvc
grpcio==1.20.1            # via tensorboard, tensorflow-gpu
h5py==2.9.0               # via keras-applications
humanize==0.5.1           # via dvc
idna==2.8                 # via requests
inflect==2.1.0            # via dvc
jsonpath-rw==1.4.0        # via dvc
keras-applications==1.0.7  # via tensorflow-gpu
keras-preprocessing==1.0.9  # via tensorflow-gpu
llvmlite==0.28.0          # via numba
markdown==3.1             # via tensorboard
mock==3.0.4               # via tensorflow-estimator
nanotime==0.5.2           # via dvc
networkx==2.3             # via dvc
numba==0.43.1             # via fastparquet
numpy==1.16.3             # via fastparquet, h5py, keras-applications, keras-preprocessing, numba, pandas, scikit-learn, scipy, tensorboard, tensorflow-estimator, tensorflow-gpu, tensorflow-hub
pandas==0.24.2
pendulum==2.0.4
pillow==6.0.0             # via asciimatics
pip-tools==3.6.1
ply==3.11                 # via dvc, jsonpath-rw
protobuf==3.7.1           # via tensorboard, tensorflow-gpu, tensorflow-hub
pyasn1==0.4.5             # via dvc
pyfiglet==0.8.post1       # via asciimatics
pyparsing==2.4.0          # via grandalf
python-dateutil==2.8.0    # via pandas, pendulum
pytz==2019.1              # via pandas
pytzdata==2019.1          # via pendulum
pyyaml==5.1               # via dvc
requests==2.21.0          # via dvc
schema==0.7.0             # via dvc
scikit-learn==0.20.3      # via sklearn
scipy==1.2.1              # via scikit-learn
six==1.12.0               # via absl-py, bert-tensorflow, configobj, fastparquet, grpcio, h5py, jsonpath-rw, keras-preprocessing, mock, pip-tools, protobuf, python-dateutil, tensorboard, tensorflow-estimator, tensorflow-gpu, tensorflow-hub, thrift
sklearn==0.0
smmap2==2.0.5             # via gitdb2
tensorboard==1.13.1       # via tensorflow-gpu
tensorflow-estimator==1.13.0  # via tensorflow-gpu
tensorflow-gpu==1.13.1
tensorflow-hub==0.4.0
termcolor==1.1.0          # via tensorflow-gpu
thrift==0.11.0            # via fastparquet
toml==0.10.0              # via black
treelib==1.5.5            # via dvc
urllib3==1.24.3           # via requests
wcwidth==0.1.7            # via asciimatics
werkzeug==0.15.2          # via tensorboard
wheel==0.33.1             # via tensorboard, tensorflow-gpu
zc.lockfile==1.4          # via dvc
